MemberListの作成 MemberのweightをDoubleからFloatに変更

+++ app/src/main/java/com/github/mutoxu_n/splitapp/models/Member.kt
@@ -7,6 +7,6 @@ import kotlinx.parcelize.Parcelize
 data class Member(
     val name: String,
     val uid: String,
-    val weight: Double,
+    val weight: Float,
     val role: Role,
 ): Parcelable
